src-agents/phase1/requirements.txt
langchain==0.2.0
langchain-community==0.2.1
langchain-core==0.2.3
langchain-openai==0.1.8
openai==1.30.5
pydantic==2.6.4
python-dotenv==1.0.1
tiktoken==0.7.0
markdown==3.6
unstructured==0.14.3
azure-search-documents==11.4.0
azure-identity==1.15.0
uvicorn==0.23.2
fastapi==0.110.0
starlette==0.37.2